Move add plate actions
