+++ ui/src/KaptureApp/actions/editorActions.js
@@ -9,7 +9,6 @@ import * as selectors from 'KaptureApp/store/editor/selectors';
 import {
   createGrid,
   createGridData,
-  createContainersForGrid,
   addContainersToGrid,
   createContainer,
   createComponent,
@@ -25,9 +24,7 @@ import {
 import { GRID_ROW_HEADERS } from 'KaptureApp/config/grid';
 
 const {
-  addGrid: _addGrid,
   addGrids: _addGrids,
-  addContainerToGrid: _addContainerToGrid,
   setGridComponents: _setGridComponents,
   // deselectGridContainers: _deselectGridContainers,
   toggleGridContainerSelections: _toggleGridContainerSelections,
@@ -51,73 +48,6 @@ export const {
 } = editorActions;
 
 const { wrapWithChangeHandler } = activity;
-
-export const addNewPlates = wrapWithChangeHandler((dimensions, quantity) => {
-  return (dispatch, getState) => {
-    const grids = [];
-    for (let i = 0; i < quantity; i++) {
-      const gridData = createGridData(dimensions, GRID_ROW_HEADERS);
-      const grid = createGrid({
-        containerType: 'Plate',
-        dimensions: dimensions,
-        data: gridData,
-      });
-      const containerPositions = createContainersForGrid(
-        dimensions,
-        'PlateWell',
-        GRID_ROW_HEADERS
-      );
-      addContainersToGrid(grid, containerPositions, GRID_ROW_HEADERS);
-      grids.push(grid);
-    }
-    dispatch(_addGrids({ grids, activeGridId: grids[0].id }));
-  };
-});
-
-export const addNewRack = wrapWithChangeHandler(({ dimensions }) => {
-  return (dispatch, getState) => {
-    const gridData = createGridData(dimensions, GRID_ROW_HEADERS);
-    const grid = createGrid({
-      containerType: 'Rack',
-      dimensions: dimensions,
-      data: gridData,
-    });
-    dispatch(_addGrid({ grid }));
-  };
-});
-
-export const addNewContainer = wrapWithChangeHandler(({ containerType }) => {
-  return (dispatch, getState) => {
-    const gridData = createGridData({ rows: 1, columns: 1 }, GRID_ROW_HEADERS);
-    const grid = createGrid({
-      containerType,
-      dimensions: { rows: 1, columns: 1 },
-      data: gridData,
-    });
-    const containerPositions = createContainersForGrid(
-      { rows: 1, columns: 1 },
-      containerType,
-      GRID_ROW_HEADERS
-    );
-    addContainersToGrid(grid, containerPositions, GRID_ROW_HEADERS);
-    dispatch(_addGrid({ grid }));
-  };
-});
-
-export const addNewContainerToGrid = wrapWithChangeHandler(
-  ({ gridId, position, containerType }) => {
-    return (dispatch, getState) => {
-      const container = createContainer({ containerType });
-      dispatch(
-        _addContainerToGrid({
-          gridId,
-          position,
-          container,
-        })
-      );
-    };
-  }
-);
 
 export const handleContainerClick = wrapWithChangeHandler(
   ({ gridId, positions }) => {
